Move DBSubcommand to a separate file
